+++ framework/sources/HFTextRepresenter.m
@@ -268,8 +268,8 @@
     if (bits & (HFControllerColorBytes)) {
         if([[self controller] shouldColorBytes]) {
             [[self view] setByteColoring: ^(uint8_t byte, uint8_t *r, uint8_t *g, uint8_t *b, uint8_t *a){
-                *r = *g = *b = 255 * ((255-byte)/255.0*0.6+0.4);
-                *a = 255 * 0.7;
+                *r = *g = *b = (uint8_t)(255 * ((255-byte)/255.0*0.6+0.4));
+                *a = (uint8_t)(255 * 0.7);
             }];
         } else {
             [[self view] setByteColoring:NULL];
